+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Postman Sandbox Changelog
 
+#### Unreleased
+* Added `pm.response.to.have.bodyWithJSONSchema()` assertion
+
 #### 2.1.3 (May 12, 2017)
 * Updated `uniscope` to v1.1.1, which contains a bugfix for handling falsey globals
 * Introduced `pm.variables` as a unified interface for variable access